requirements.txt
torchaudio
soundfile
plotly
tqdm